add/verification api

+++ app/src/main/java/com/example/planup/main/goal/item/GoalItem.kt
@@ -1,7 +1,7 @@
 package com.example.planup.main.goal.item
 
 data class GoalItem(
-    val goalId: Long,
+    val goalId: Int,
     val title: String,
     val description: String,
     val percent: Int,
